# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Tour Space app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of Tour Space api"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    },
    {
      "url": "https://tour-space-api.onrender.com"
    }
  ],
...
  "paths": {
...
    "/reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Get all reviews",
        "operationId": "getReviews",
        "description": "Get all reviews",
        "parameters": [
          {
            "in": "query",
            "name": "sortOrder",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "example": "asc"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "reviews",
                    "totalPages",
                    "page",
                    "hasNextPage"
                  ],
                  "properties": {
                    "reviews": {
                      "type": "array"
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 5
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "hasNextPage": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Add review about tour",
        "operationId": "addReview",
        "description": "Add review about tour",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userName",
                  "tourCountry",
                  "review",
                  "rating"
                ],
                "properties": {
                  "userName": {
                    "description": "User's name",
                    "type": "string",
                    "example": "Olha Ivanenko"
                  },
                  "tourCountry": {
                    "description": "Tour's country",
                    "type": "string",
                    "example": "Італія"
                  },
                  "review": {
                    "description": "Review about tour",
                    "type": "string",
                    "example": "Це була чудова подорож, яку я рекомендуватису всім знайомим!"
                  },
                  "rating": {
                    "description": "Tour rating",
                    "type": "number",
                    "example": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Adds review about tour",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Review added"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "400": {
        "description": "Bad request error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "message",
                "status"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Bad request"
                },
                "status": {
                  "type": "integer",
                  "example": 400
                }
              }
            }
          }
        }
      },
...
    }
  }
}